% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch.R
\name{big_parallelize}
\alias{big_parallelize}
\title{big_parallelize}
\usage{
big_parallelize(
  X,
  p.FUN,
  p.combine = NULL,
  ind = bigstatsr::cols_along(X),
  ncores = 4,
  ...
)
}
\arguments{
\item{X}{is a file-backed data matrix}

\item{p.FUN}{is a function to apply}

\item{p.combine}{is a rule to combine the results from each core}

\item{ind}{is a vector of column indeces on which the computations are performed}

\item{ncores}{is a nuber of cores}

\item{...}{are additional parameters}
}
\value{
The result of the parallel computations
}
\description{
parallel call a function on file-backed matrix
}
\keyword{internal}
