% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{chickn}
\alias{chickn}
\title{chickn-package}
\description{
R package \code{chickn} implements Chromatogram Hierarchical Compressive K-means with Nystrom approximation
clustering approach. It is designed to cluster a large collection of high-resolution
mass spectrometry signals (chromatographic profiles) relying on the compressed data version (data sketch).
Data compression is achieved using the Nystrom method and the sketch operator from \insertCite{DBLP:journals/corr/KerivenBGP16}{chickn}.
The Filebacked Big Matrix (FBM) class from the \href{https://github.com/privefl/bigstatsr}{bigstatsr} package
is used to store and to manupulate matrices, which cannot be load in memory.
}
\references{
\itemize{
\item \insertRef{wang2019scalable}{chickn}
\item \insertRef{DBLP:journals/corr/KerivenBGP16}{chickn}.
}
}
\author{
Olga Permiakova
}
