% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_Mat}
\alias{cumsum_Mat}
\title{cumsum_Mat}
\usage{
cumsum_Mat(C)
}
\arguments{
\item{C}{is a numeric matrix}
}
\value{
A matrix of cummulative sum
}
\description{
rcpp cummulative sum of matrix columns
}
\keyword{internal}
