% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_parallel}
\alias{cumsum_parallel}
\title{Cumulative sum computation}
\usage{
cumsum_parallel(X, A_cumsum)
}
\arguments{
\item{X}{A Filebacked Big Matrix n x N.}

\item{A_cumsum}{A Filebacked Big Matrix n x N, where cummulative sums are stored.}
}
\description{
Parallel implementation of the cumulative sum of the matrix columns.
}
