% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_sug}
\alias{cumsum_sug}
\title{cumsum_sug}
\usage{
cumsum_sug(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
A vector of cummulative sum
}
\description{
rcpp cummulative sum of integer vector
}
\keyword{internal}
