% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_funR.R
\name{DB_funR}
\alias{DB_funR}
\title{DB_funR}
\usage{
DB_funR(Data_cumsum, Cluster_partition, C, DIR_output, ncores = 4)
}
\arguments{
\item{Data_cumsum}{A Filebacked Big Matrix with signal cumultive sums in columns}

\item{Cluster_partition}{A list of the resulting cluster assignment. Each list element correspond to a cluster with a list of data vector indices.}

\item{C}{A matrix of cluster centroids.}

\item{DIR_output}{An output directory for temporal files.}

\item{ncores}{Number of cores. Default is 4.}
}
\value{
A positive scalar.
}
\description{
Davies Bouldin index computation.
}
\keyword{internal}
