% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DBindex}
\alias{DBindex}
\title{DBindex}
\usage{
DBindex(X_cumsum, rowInd, Clusters, C, cluster_size, DistInter)
}
\arguments{
\item{X_cumsum}{is a Filebacked Big Matrix, which contains column matrix cumulative sums.}

\item{rowInd}{is a vector of the row indeces}

\item{Clusters}{is a list of cluster assignment}

\item{C}{is a matrix of the cluster centroids}

\item{cluster_size}{is a vector of the cluster sizes}

\item{DistInter}{is a Filebacked Big Matrix, which stores distances between a cluster centoid and signals in the cluster}
}
\value{
A value of the DB index
}
\description{
Davies Bouldin index computation.
}
\keyword{internal}
