% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DBindex_BIG}
\alias{DBindex_BIG}
\title{DBindex_BIG}
\usage{
DBindex_BIG(
  X_cumsum,
  rowInd,
  Clusters,
  C,
  C_cumsum,
  cluster_size,
  DistInter,
  R
)
}
\arguments{
\item{X_cumsum}{is a Filebacked data cummulative sum matrix n x N}

\item{rowInd}{is a vector of the row indeces}

\item{Clusters}{is a list of cluster assignment}

\item{C}{is a matrix of the cluster centroids}

\item{C_cumsum}{is a matrix of the cummulative sums of cluster centroids}

\item{cluster_size}{is a vector of the cluster sizes}

\item{DistInter}{is a matrix of the inter distances}

\item{R}{is a matrix of the intra distances}
}
\value{
A value of the DB index
}
\description{
Davies Bouldin index computation
}
\keyword{internal}
