% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_parallel}
\alias{E_parallel}
\title{Euclidean distance}
\usage{
E_parallel(X, C, set_c)
}
\arguments{
\item{X}{A Filebacked Big Matrix \eqn{n \times N}.}

\item{C}{A Filebacked Big Matrix \eqn{N \times l}, which stores the Euclidean distances.}

\item{set_c}{Column index vector. The data vector indices for which the Euclidean distances are computed.}
}
\description{
Euclidean distance
}
\details{
The Euclidean distances are computed between the data vectors from \code{set_c} and all columns of X.
}
