% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_fun_cpp}
\alias{G_fun_cpp}
\title{G_fun_cpp}
\usage{
G_fun_cpp(x, y, W)
}
\arguments{
\item{x}{is a vector}

\item{y}{is a vector}

\item{W}{is a frequency matrix}
}
\value{
A vector
}
\description{
Function for gradient computation.
}
\keyword{internal}
