% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_estimation}
\alias{gamma_estimation}
\title{Kernel parameter estimation}
\usage{
gamma_estimation(X, size, kernel_type)
}
\arguments{
\item{X}{A Filebacked Big Matrix \eqn{n \times N}.}

\item{size}{Neighborhood size.}

\item{kernel_type}{Kernel function type. Available types are c("Gaussian", "Laplacian").}
}
\value{
The estimated kernel parameter.
}
\description{
Kernel parameter estimation by averaging the distances to the closest neighbors.
}
