% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gradient_OMP_cpp}
\alias{Gradient_OMP_cpp}
\title{Gradient OMP}
\usage{
Gradient_OMP_cpp(c, W, residue)
}
\arguments{
\item{c}{is a cluster centroid}

\item{W}{is a frequency matrix mxs}

\item{residue}{is a vector}
}
\value{
The gradient vector
}
\description{
Gradient with respect to cluster centroid vector
}
\keyword{internal}
