% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InterDist}
\alias{InterDist}
\title{Computation of the Wasserstein distances between cluster centroids}
\usage{
InterDist(C_cumsum)
}
\arguments{
\item{C_cumsum}{is a Filebacked data cummulative sum matrix of N cluster centroids}
}
\value{
N x N matrix of the Wasserstein distances
}
\description{
Computation of the Wasserstein distances between cluster centroids
}
\keyword{internal}
