% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IntraDist}
\alias{IntraDist}
\title{Computation of the Wasserstein distances between cluster centroid and signals in the cluster}
\usage{
IntraDist(X_cumsum, rowInd, colInd, C_cumsum)
}
\arguments{
\item{X_cumsum}{is a Filebacked data cummulative sum matrix n x N}

\item{rowInd}{is a vector of the row indeces}

\item{colInd}{is a vector of the column indeces}

\item{C_cumsum}{is a vector of the cummulative sum of the cluster centroid}
}
\value{
The vector of the Wasserstein distances
}
\description{
Computation of the Wasserstein distances between cluster centroid and signals in the cluster
}
\keyword{internal}
