% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preimage_average.R
\name{Preimage}
\alias{Preimage}
\title{Preimage}
\usage{
Preimage(
  Data,
  K_W1,
  C_out,
  Cl_assign,
  max_Nsize = 32,
  ncores = 4,
  DIR_out = getwd(),
  BIG = FALSE
)
}
\arguments{
\item{Data}{A Filebacked Big Matrix n x N. Data signals are stored in the matrix columns.}

\item{K_W1}{A Filebacked Big Matrix of the Nystrom kernel matrix \eqn{s \times N},
where N is the number of signal in the training collection and s is the Nystrom sample size.}

\item{C_out}{A Filebacked Big Matrix of cluster centroids.}

\item{Cl_assign}{A Filebacked Big matrix of the cluster assignment.}

\item{max_Nsize}{Maxumum number of neighbors.}

\item{ncores}{Number of cores.}

\item{DIR_out}{A directory to save the result, by default it is the working directory.}

\item{BIG}{logical parameter that controls whether the resulting consensus chromatograms are stored as a Filebacked Big Matrix ('Centroid_preimage.bk').
Default is FALSE.}
}
\value{
A matrix or a Filebacked Big Matrix if \code{BIG} = TRUE.
}
\description{
Consensus chromatogram computation.
}
