% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RandIndex}
\alias{RandIndex}
\title{RandIndex}
\usage{
RandIndex(Data)
}
\arguments{
\item{Data}{is a two columns matrix. First column contains the clustering assignment result,
second column ground truth clustering assignment each row correspond to the data vector index.}
}
\value{
The rand index value
}
\description{
Rand Index computation
}
\keyword{internal}
