% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch.R
\name{split_index}
\alias{split_index}
\title{split_index}
\usage{
split_index(total_len, block_len, nb_split = ceiling(total_len/block_len))
}
\arguments{
\item{total_len}{is a total length of the index vector}

\item{block_len}{is a length of the block}

\item{nb_split}{is a number of the splits}
}
\value{
A list of the splits that contains lower, upper block indeces and block size
}
\description{
splits the indeces into several blocks
}
\keyword{internal}
