% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPS2}
\alias{UPS2}
\title{UPS2 dataset}
\format{
A matrix with 1653 rows and 190 columns.
}
\source{
\url{https://github.com/optimusmoose/ups2GT}
}
\usage{
data(UPS2)
}
\description{
Proteomics data acquired within the mass spectrometry analysis of UPS2 sample.
}
\details{
Only a small part of data was taken from
the original dataset described in \insertCite{henning2018peptide}{chickn}.
The UPS2 dataset contains 190 chromatographic traces (matrix columns) acquired along
the retention time (matrix rows) in the liquid chromatography.
}
\references{
\itemize{
\item \insertRef{tsou2015dia}{chickn}
\item \insertRef{henning2018peptide}{chickn}
}
}
\keyword{datasets}
