% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W1_centr_centr}
\alias{W1_centr_centr}
\title{Parallel computation of the Wasserstein distances between cluster centroids}
\usage{
W1_centr_centr(C_cumsum, Dist)
}
\arguments{
\item{C_cumsum}{is a Filebacked data cummulative sum matrix of N cluster centroids}

\item{Dist}{is a Filebacked inter distance matrix}
}
\value{
N x N matrix of the Wasserstein distances
}
\description{
Parallel computation of the Wasserstein distances between cluster centroids
}
\keyword{internal}
