% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W1_centr_centr_BIG}
\alias{W1_centr_centr_BIG}
\title{Parallel computation for big clusters of the Wasserstein distances between cluster centroids}
\usage{
W1_centr_centr_BIG(C_cumsum, Dist)
}
\arguments{
\item{C_cumsum}{is a Filebacked data cummulative sum matrix n x N}

\item{Dist}{is a Filebacked matrix of the Wasserstein distances between centroids}
}
\description{
Parallel computation for big clusters of the Wasserstein distances between cluster centroids
}
\keyword{internal}
