% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W1_cl_centr_BIG}
\alias{W1_cl_centr_BIG}
\title{W1_cl_centr_BIG}
\usage{
W1_cl_centr_BIG(X_cumsum, rowInd, colInd, C_cumsum, nbr_cluster)
}
\arguments{
\item{X_cumsum}{is a Filebacked data cummulative sum matrix n x N}

\item{rowInd}{is a vector of the row indeces}

\item{colInd}{is a vector of the column indeces}

\item{C_cumsum}{is a vector of the cummulative sum of the cluster centroid}

\item{nbr_cluster}{is a number of clusters}
}
\value{
A vector of the Wasserstein distances
}
\description{
Parallel computation for big clusters of
the Wasserstein distances between cluster centroid and signals in the cluster
}
\keyword{internal}
