% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W1_parallel}
\alias{W1_parallel}
\title{Wasserstein-1 distance}
\usage{
W1_parallel(X, C, set_c)
}
\arguments{
\item{X}{A Filebacked Big Matrix n x N.}

\item{C}{A Filebacked Big Matrix N x l, which stores the Wasserstein distances.}

\item{set_c}{Column index vector. The data vector indices for which the Wasserstein distances are computed.}
}
\description{
Wasserstein-1 distance
}
\details{
The Wasserstein-1 distances are computed between the data vectors from \code{set_c} and all columns of X.
}
