% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{chickn}
\alias{chickn}
\title{chickn-package}
\description{
The R package \code{chickn} implements the Chromatogram Hierarchical Compressive K-means with Nystrom approximation
clustering approach. It is designed to cluster a large collection of high-resolution
mass spectrometry signals (chromatographic profiles) relying on a compressed version of the data (a.k.a. data sketch).
Data compression is achieved following the guidelines for Nystrom approximation provided by \insertCite{wang2019scalable}{chickn}
and the sketching operator from \insertCite{DBLP:journals/corr/KerivenBGP16}{chickn}.
The Filebacked Big Matrix (FBM) class from the \href{https://github.com/privefl/bigstatsr}{bigstatsr} package
is used to store and to manupulate matrices, which cannot be load in memory.
}
\references{
\itemize{
\item Permiakova O, Guibert R, Kraut A, Fortin T, Hesse AM, Burger T (2020) "CHICKN: Extraction of peptide chromatographic
elution profiles from large scale mass spectrometry data by means of Wasserstein compressive hierarchical cluster analysis."
BMC Bioinformatics (under revision).
}
}
\author{
Olga Permiakova, Romain Guibert, Thomas Burger
}
